Rename files to reflect actual dataset to convert
