+++ src/schwerdt_lab_to_nwb/amjad_2025/conversion_notes.md
