--- src/schwerdt_lab_to_nwb/choi_2025/metadata.yaml
@@ -1,20 +1,0 @@
-NWBFile:
-  keywords:
-    - Keyword1
-    - Keyword2
-    - Keyword3
-  related_publications:
-    https://doi.org/### or link to APA or MLA citation of the publication
-  session_description:
-    A rich text description of the experiment. Can also just be the abstract of the publication.
-  institution: Institution where the lab is located
-  lab: Schwerdt
-  experimenter:
-    - Last, First Middle
-    - Last, First Middle
-Subject:
-  species: Rattus norvegicus
-  description: A rich text description of the subject
-  age: TBD  # in ISO 8601, such as "P1W2D"
-  sex: TBD  # One of M, F, U, or O
-  date_of_birth: 2014-06-22 00:00:00-04:00  # Example
